pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
flask = "^3.0.2"
flask-sqlalchemy = "^3.1.1"
uuid = "^1.30"
passlib = "^1.7.4"
marshmallow = "^3.21.1"
flask-smorest = "^0.44.0"
flask-migrate = "^4.0.7"
sqlalchemy = "^2.0.28"
mysql-connector-python = "^8.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
